+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The profile view requests the latest events and metadata for the given user from the relays
 - Add the ellipsis button to NoteCard and allow the user to copy the NIP-19 note ID of a note.
 - Enabled button to copy user ID on Profile View
+- Add PostHog analytics
 
 ## [0.1 (5)] 2023-03-02
