+++ Nos/Views/Onboarding/OnboardingView.swift
@@ -8,27 +8,6 @@
 import SwiftUI
 import Dependencies
 
-struct ViewDidLoadModifier: ViewModifier {
-    @State private var viewDidLoad = false
-    let action: (() -> Void)?
-    
-    func body(content: Content) -> some View {
-        content
-            .onAppear {
-                if viewDidLoad == false {
-                    viewDidLoad = true
-                    action?()
-                }
-            }
-    }
-}
-
-extension View {
-    func onViewDidLoad(perform action: (() -> Void)? = nil) -> some View {
-        self.modifier(ViewDidLoadModifier(action: action))
-    }
-}
-
 struct OnboardingView: View {
     enum OnboardingStep {
         case getStarted
